+++ src/Serializer/PayloadSerializer.php
@@ -9,7 +9,10 @@ use Sentry\Event;
 use Sentry\EventType;
 use Sentry\ExceptionDataBag;
 use Sentry\Frame;
+use Sentry\Options;
+use Sentry\Tracing\DynamicSamplingContext;
 use Sentry\Tracing\Span;
+use Sentry\Tracing\TransactionMetadata;
 use Sentry\Util\JSON;
 
 /**
@@ -20,6 +23,16 @@ use Sentry\Util\JSON;
  */
 final class PayloadSerializer implements PayloadSerializerInterface
 {
+    /**
+     * @var Options The SDK client options
+     */
+    private $options;
+
+    public function __construct(Options $options)
+    {
+        $this->options = $options;
+    }
+
     /**
      * {@inheritdoc}
      */
@@ -106,6 +119,7 @@ final class PayloadSerializer implements PayloadSerializerInterface
                 'username' => $user->getUsername(),
                 'email' => $user->getEmail(),
                 'ip_address' => $user->getIpAddress(),
+                'segment' => $user->getSegment(),
             ]);
         }
 
@@ -160,6 +174,11 @@ final class PayloadSerializer implements PayloadSerializerInterface
 
         if (EventType::transaction() === $event->getType()) {
             $result['spans'] = array_values(array_map([$this, 'serializeSpan'], $event->getSpans()));
+            $transactionMetadata = $event->getSdkMetadata('transaction_metadata');
+
+            if ($transactionMetadata instanceof TransactionMetadata) {
+                $result['transaction_info']['source'] = (string) $transactionMetadata->getSource();
+            }
         }
 
         $stacktrace = $event->getStacktrace();
@@ -175,17 +194,33 @@ final class PayloadSerializer implements PayloadSerializerInterface
 
     private function serializeAsEnvelope(Event $event): string
     {
-        $envelopeHeader = JSON::encode([
+        // @see https://develop.sentry.dev/sdk/envelopes/#envelope-headers
+        $envelopeHeader = [
             'event_id' => (string) $event->getId(),
             'sent_at' => gmdate('Y-m-d\TH:i:s\Z'),
-        ]);
+            'dsn' => (string) $this->options->getDsn(),
+            'sdk' => [
+                'name' => $event->getSdkIdentifier(),
+                'version' => $event->getSdkVersion(),
+            ],
+        ];
 
-        $itemHeader = JSON::encode([
+        $dynamicSamplingContext = $event->getSdkMetadata('dynamic_sampling_context');
+
+        if ($dynamicSamplingContext instanceof DynamicSamplingContext) {
+            $entries = $dynamicSamplingContext->getEntries();
+
+            if (!empty($entries)) {
+                $envelopeHeader['trace'] = $entries;
+            }
+        }
+
+        $itemHeader = [
             'type' => (string) $event->getType(),
             'content_type' => 'application/json',
-        ]);
+        ];
 
-        return sprintf("%s\n%s\n%s", $envelopeHeader, $itemHeader, $this->serializeAsEvent($event));
+        return sprintf("%s\n%s\n%s", JSON::encode($envelopeHeader), JSON::encode($itemHeader), $this->serializeAsEvent($event));
     }
 
     /**