+++ CHANGELOG.md
@@ -2,7 +2,11 @@
 
 ## Unreleased
 
-- Expose a function to retrieve the URL of the CSP endpoint (#1378)
+- feat: Expose a function to retrieve the URL of the CSP endpoint (#1378)
+- feat: Add support for Dynamic Sampling (#1360)
+  - Add `segment` to `UserDataBag`
+  - Add `TransactionSource`, to set information about the transaction name
+  - Deprecate `TransactionContext::fromSentryTrace()` in favor of `TransactionContext::fromHeaders()`
 
 ## 3.8.1 (2022-09-21)